Cyber_Owl_Mobile_App/bridge_backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
websockets
pydantic
argon2-cffi
pymongo
motor
dnspython
openai
